+++ libs/core/notification/directives/notification-popover.directive.ts
@@ -1,0 +1,10 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    selector: '[fdNotificationPopover], [fd-notification-popover]',
+    standalone: true,
+    host: {
+        class: 'fd-notification-popover'
+    }
+})
+export class NotificationPopoverDirective {}