+++ libs/core/notification/token.ts
@@ -1,0 +1,8 @@
+import { InjectionToken } from '@angular/core';
+
+export const FD_NOTIFICATION = new InjectionToken('FdNotificationComponent');
+export const FD_NOTIFICATION_TITLE = new InjectionToken('FdNotificationTitleDirective');
+export const FD_NOTIFICATION_PARAGRAPH = new InjectionToken('FdNotificationParagraphDirective');
+export const FD_NOTIFICATION_GROUP_HEADER = new InjectionToken('FdNotificationGroupHeaderComponent');
+export const FD_NOTIFICATION_GROUP_HEADER_TITLE = new InjectionToken('FdNotificationGroupHeaderTitleDirective');
+export const FD_NOTIFICATION_GROUP_LIST = new InjectionToken('FdNotificationGroupListComponent');