+++ libs/docs/core/notification/examples/component-as-content/notification-component-as-content-example.component.ts
@@ -28,8 +28,7 @@ export class NotificationComponentAsContentExampleComponent {
                 title: 'Notification Title',
                 paragraph: 'Notification Description',
                 firstFooterContent: 'SAP Analytics Cloud',
-                secondFooterContent: 'Just Now',
-                open: 'Open'
+                secondFooterContent: 'Just Now'
             },
             width: '500px'
         });